feat(home, agents): add status for agents, timer intervention and add agent in the home
