+++ src/app/app.module.ts
@@ -8,7 +8,7 @@ import { StationComponent } from "./components/station/station.component";
 import {StationPreviewComponent } from "./screens/station-preview/station-preview.component";
 import {AppRoutingModule} from "./app-routing.module";
 import {AppService} from "../services/app.service";
-import { AgentsComponent } from './agents/agents.component';
+import { AgentsComponent } from './components/agents/agents.component';
 
 @NgModule({
   declarations: [AppComponent, HomeComponent, StationComponent, StationPreviewComponent, AgentsComponent],
